+++ poche/app/src/main/java/com/timdeve/poche/model/Feed.kt
@@ -4,30 +4,30 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Feed(
-    val id: Int,
+    val id: Long,
     val name: String,
     val url: String,
 )
 
 @Serializable
 data class FeedList(
-    val id: Int,
+    val id: Long,
     val name: String,
-    val feedIds: List<Int>,
+    val feedIds: List<Long>,
 )
 
 // genFeeds generates fake data for previews
-fun genFeeds(): Pair<Map<Int, Feed>, Map<Int, FeedList>> {
+fun genFeeds(): Pair<Map<Long, Feed>, Map<Long, FeedList>> {
     return Pair(
         listOf(
-            0 to Feed(0, "Hacker News", "http://example.com"),
-            1 to Feed(1, "News Org", "http://example.com"),
-            2 to Feed(2, "Tech Knowledge", "http://example.com"),
-            3 to Feed(3, "Stuff N' Things", "http://example.com"),
+            0L to Feed(0L, "Hacker News", "http://example.com"),
+            1L to Feed(1L, "News Org", "http://example.com"),
+            2L to Feed(2L, "Tech Knowledge", "http://example.com"),
+            3L to Feed(3L, "Stuff N' Things", "http://example.com"),
         ).toMap(),
         listOf(
-            0 to FeedList(0, "My List", listOf(0, 3)),
-            1 to FeedList(1, "Another one", listOf(0, 3)),
+            0L to FeedList(0L, "My List", listOf(0, 3)),
+            1L to FeedList(1L, "Another one", listOf(0, 3)),
         ).toMap()
     )
 }
