+++ poche/app/src/main/java/com/timdeve/poche/ui/screens/stories/StoriesRoute.kt
@@ -13,7 +13,7 @@ fun StoriesRoute(
     storiesViewModel: StoriesViewModel,
     feedsViewModel: FeedsViewModel,
     navController: NavHostController,
-    listId: Int? = null,
+    listId: Long? = null,
 ) {
     LaunchedEffect(Unit) {
         storiesViewModel.setListId(listId)
